alterações feitas até 26-11-23

+++ aula29.py
@@ -7,14 +7,14 @@ numero_str = input('Vou dobrar o número que você digitar: ')
 
 try:
     print('STR: ', numero_str)
-    numero_float = int(numero_str)
+    numero_float = float(numero_str)
     print('FLOAT:', numero_float)
     print(f'O dobro de {numero_str} é {numero_float * 2:.2f}')
 except:
     print('Isso não é um número')
 
 if numero_str.isdigit():
-    numero_float = int(numero_str)
+    numero_float = float(numero_str)
     print(f'O dobro de {numero_str} é {numero_float * 2:.2f}')
 else:
     print('Isso não é um número')
